test(mxc-sdk): move the Windows streaming test out of the macOS-gated file
`tests/streaming.rs` is `#![cfg(target_os = "macos")]`, so the
`#[cfg(target_os = "windows")]` ProcessContainer streaming test it contained
could never compile -- the intended Windows coverage was silently missing.
Move that test into its own `tests/streaming_processcontainer.rs`, gated
`#![cfg(target_os = "windows")]`, so it actually builds on Windows.

Addresses a Copilot review comment on #556.
