+++ src/core/mxc-sdk/tests/streaming.rs
@@ -196,47 +196,6 @@ fn streaming_stdout_closer_unblocks_parked_read_without_killing() {
     let _ = proc.wait();
 }
 
-// ---------------------------------------------------------------------------
-// Windows ProcessContainer streaming — integration test. Requires an elevated,
-// host-prepped Windows host (see docs/host-prep.md), so it is `#[ignore]`d.
-// ---------------------------------------------------------------------------
-
-#[cfg(target_os = "windows")]
-#[test]
-#[ignore = "requires an elevated, host-prepped Windows host (see docs/host-prep.md)"]
-fn streaming_processcontainer_bidirectional_stdio() {
-    use std::io::{Read, Write};
-
-    let policy = SandboxPolicy {
-        version: "0.7.0-alpha".to_string(),
-        filesystem: Some(mxc_sdk::policy::FilesystemSection {
-            readwrite_paths: vec!["C:\\Windows\\Temp".to_string()],
-            readonly_paths: vec![],
-            denied_paths: vec![],
-            clear_policy_on_exit: None,
-        }),
-        network: None,
-        ui: None,
-        timeout_ms: None,
-    };
-    let mut request = build_request(&policy, None).expect("build_request");
-    // `cmd /c more` echoes stdin to stdout until EOF, then exits.
-    request.set_script("cmd /c more");
-    let mut proc = spawn_sandbox(request).expect("spawn");
-
-    let mut stdin = proc.take_stdin().expect("stdin available");
-    let mut stdout = proc.take_stdout().expect("stdout available");
-
-    stdin.write_all(b"ping-pong\r\n").expect("write stdin");
-    drop(stdin);
-
-    let mut out = String::new();
-    stdout.read_to_string(&mut out).expect("read stdout");
-    assert!(out.contains("ping-pong"), "got: {:?}", out);
-
-    assert_eq!(proc.wait().expect("wait"), WaitOutcome::Exited(0));
-}
-
 #[cfg(target_os = "macos")]
 #[test]
 fn streaming_wait_discards_untaken_streams() {
